Code/requirements.txt
pandas
awscli
pyodbc
fastapi
uvicorn[standard] 
pydantic 
azure-storage-blob
sqlalchemy 
python-dotenv


